Add Models, Modify db, Add queries, Edit login and register routes
